+++ integration_tests/pytest.ini
@@ -38,6 +38,7 @@ markers =
     pyarrow_test: Mark pyarrow tests
     datagen_overrides: Mark that allows overriding datagen settings (i.e. seed) for a test
     tz_sensitive_test: Mark the test as a time zone sensitive test which will be tested against extra timezones
+    protobuf_test: Mark a test that requires explicit spark-protobuf jars
     spark_job_timeout(seconds, dump_threads): Override the default timeout setting
 filterwarnings =
     ignore:.*pytest.mark.order.*:_pytest.warning_types.PytestUnknownMarkWarning
