+++ integration_tests/run_pyspark_from_build.sh
@@ -46,8 +46,9 @@
 #   To run all tests, including Avro tests:
 #     INCLUDE_SPARK_AVRO_JAR=true ./run_pyspark_from_build.sh
 #
-#   To run tests WITHOUT Protobuf tests (protobuf is included by default):
-#     INCLUDE_SPARK_PROTOBUF_JAR=false ./run_pyspark_from_build.sh
+#   To run protobuf tests with explicit jars:
+#     PROTOBUF_JARS=/path/to/spark-protobuf.jar,/path/to/protobuf-java.jar \
+#       ./run_pyspark_from_build.sh -m protobuf_test
 #
 #   To run a specific test:
 #     TEST=my_test ./run_pyspark_from_build.sh
@@ -144,101 +145,40 @@ else
         AVRO_JARS=""
     fi
 
-    # Protobuf support: Include spark-protobuf jar by default for protobuf_test.py
-    # Set INCLUDE_SPARK_PROTOBUF_JAR=false to disable
-    PROTOBUF_JARS=""
-    if [[ $( echo ${INCLUDE_SPARK_PROTOBUF_JAR} | tr '[:upper:]' '[:lower:]' ) != "false" ]];
+    # Protobuf tests require explicit extra jars, similar to hybrid execution.
+    PROTOBUF_EXTRA_JARS=""
+    export PROTOBUF_JARS_AVAILABLE=false
+    if [[ -n "${PROTOBUF_JARS}" ]];
     then
-        export INCLUDE_SPARK_PROTOBUF_JAR=true
-        mkdir -p "${TARGET_DIR}/dependency"
-        
-        # Download spark-protobuf jar if not already in target/dependency
-        PROTOBUF_JAR_NAME="spark-protobuf_${SCALA_VERSION}-${VERSION_STRING}.jar"
-        PROTOBUF_JAR_PATH="${TARGET_DIR}/dependency/${PROTOBUF_JAR_NAME}"
-        
-        if [[ ! -f "$PROTOBUF_JAR_PATH" ]]; then
-            echo "Downloading spark-protobuf jar..."
-            PROTOBUF_MAVEN_URL="https://repo1.maven.org/maven2/org/apache/spark/spark-protobuf_${SCALA_VERSION}/${VERSION_STRING}/${PROTOBUF_JAR_NAME}"
-            if curl -fsL -o "$PROTOBUF_JAR_PATH" "$PROTOBUF_MAVEN_URL"; then
-                echo "Downloaded spark-protobuf jar to $PROTOBUF_JAR_PATH"
-            else
-                echo "WARNING: Failed to download spark-protobuf jar from $PROTOBUF_MAVEN_URL"
-                rm -f "$PROTOBUF_JAR_PATH"
-            fi
-        fi
-        
-        # Also download protobuf-java jar (required dependency).
-        # Detect version from the jar bundled with Spark, fall back to version mapping.
-        PROTOBUF_JAVA_VERSION=""
-        BUNDLED_PB_JAR=$(ls "$SPARK_HOME"/jars/protobuf-java-[0-9]*.jar 2>/dev/null | sort -V | tail -1)
-        if [[ -n "$BUNDLED_PB_JAR" ]]; then
-            PROTOBUF_JAVA_VERSION=$(basename "$BUNDLED_PB_JAR" | sed 's/protobuf-java-\(.*\)\.jar/\1/')
-            echo "Detected protobuf-java version $PROTOBUF_JAVA_VERSION from SPARK_HOME"
-        fi
-        if [[ -z "$PROTOBUF_JAVA_VERSION" ]]; then
-            case "$VERSION_STRING" in
-                3.4.*) PROTOBUF_JAVA_VERSION="3.25.1" ;;
-                3.5.*) PROTOBUF_JAVA_VERSION="3.25.1" ;;
-                4.0.*|4.1.*) PROTOBUF_JAVA_VERSION="4.29.3" ;;
-                *)     PROTOBUF_JAVA_VERSION="3.25.1" ;;
-            esac
-            echo "Using protobuf-java version $PROTOBUF_JAVA_VERSION based on Spark $VERSION_STRING"
-        fi
-        PROTOBUF_JAVA_JAR_NAME="protobuf-java-${PROTOBUF_JAVA_VERSION}.jar"
-        PROTOBUF_JAVA_JAR_PATH="${TARGET_DIR}/dependency/${PROTOBUF_JAVA_JAR_NAME}"
-        
-        if [[ ! -f "$PROTOBUF_JAVA_JAR_PATH" ]]; then
-            echo "Downloading protobuf-java jar..."
-            PROTOBUF_JAVA_MAVEN_URL="https://repo1.maven.org/maven2/com/google/protobuf/protobuf-java/${PROTOBUF_JAVA_VERSION}/${PROTOBUF_JAVA_JAR_NAME}"
-            if curl -fsL -o "$PROTOBUF_JAVA_JAR_PATH" "$PROTOBUF_JAVA_MAVEN_URL"; then
-                echo "Downloaded protobuf-java jar to $PROTOBUF_JAVA_JAR_PATH"
-            else
-                echo "WARNING: Failed to download protobuf-java jar from $PROTOBUF_JAVA_MAVEN_URL"
-                rm -f "$PROTOBUF_JAVA_JAR_PATH"
-            fi
-        fi
-        
         SPARK_PROTOBUF_JAR_AVAILABLE=false
-        PROTOBUF_JAVA_AVAILABLE=false
+        VALID_PROTOBUF_JARS=()
+        IFS=',:' read -ra RAW_PROTOBUF_JARS <<< "${PROTOBUF_JARS}"
+        for raw_pb_jar in "${RAW_PROTOBUF_JARS[@]}"; do
+            pb_jar=$(readlink -e "$raw_pb_jar" || true)
+            if [[ -z "$pb_jar" ]]; then
+                echo "WARNING: Protobuf jar not found: $raw_pb_jar"
+                continue
+            fi
+            VALID_PROTOBUF_JARS+=("$pb_jar")
+            if [[ "$(basename "$pb_jar")" == spark-protobuf_* ]]; then
+                SPARK_PROTOBUF_JAR_AVAILABLE=true
+            fi
+        done
 
-        if [[ -f "$PROTOBUF_JAR_PATH" ]]; then
-            PROTOBUF_JARS="$PROTOBUF_JAR_PATH"
-            echo "Including spark-protobuf jar: $PROTOBUF_JAR_PATH"
-            SPARK_PROTOBUF_JAR_AVAILABLE=true
-        fi
-        if [[ -f "$PROTOBUF_JAVA_JAR_PATH" ]]; then
-            PROTOBUF_JARS="${PROTOBUF_JARS:+$PROTOBUF_JARS }$PROTOBUF_JAVA_JAR_PATH"
-            echo "Including protobuf-java jar: $PROTOBUF_JAVA_JAR_PATH"
-            PROTOBUF_JAVA_AVAILABLE=true
-        elif [[ -n "$BUNDLED_PB_JAR" ]]; then
-            echo "Using bundled protobuf-java jar from SPARK_HOME: $BUNDLED_PB_JAR"
-            PROTOBUF_JAVA_AVAILABLE=true
-        fi
-
-        if [[ "$SPARK_PROTOBUF_JAR_AVAILABLE" == "true" && \
-              "$PROTOBUF_JAVA_AVAILABLE" == "true" ]]; then
+        if [[ ${#VALID_PROTOBUF_JARS[@]} -gt 0 && \
+              "$SPARK_PROTOBUF_JAR_AVAILABLE" == "true" ]]; then
+            PROTOBUF_EXTRA_JARS="${VALID_PROTOBUF_JARS[*]}"
             export PROTOBUF_JARS_AVAILABLE=true
+            echo "Including protobuf jars: ${VALID_PROTOBUF_JARS[*]}"
         else
-            echo "WARNING: Protobuf JAR dependencies incomplete; protobuf tests will be skipped"
-            echo "  spark-protobuf available: $SPARK_PROTOBUF_JAR_AVAILABLE"
-            echo "  protobuf-java available: $PROTOBUF_JAVA_AVAILABLE"
-            export PROTOBUF_JARS_AVAILABLE=false
+            echo "WARNING: Protobuf jars are not configured correctly; protobuf tests will be skipped"
+            echo "  Expected PROTOBUF_JARS to include at least one spark-protobuf jar"
         fi
-        # Also add protobuf jars to driver classpath for Class.forName() to work
-        # This is needed because --jars only adds to executor classpath
-        if [[ -n "$PROTOBUF_JARS" ]]; then
-            PROTOBUF_DRIVER_CP=$(echo "$PROTOBUF_JARS" | tr ' ' ':')
-            export PYSP_TEST_spark_driver_extraClassPath="${PYSP_TEST_spark_driver_extraClassPath:+${PYSP_TEST_spark_driver_extraClassPath}:}${PROTOBUF_DRIVER_CP}"
-            echo "Added protobuf jars to driver classpath"
-        fi
-    else
-        export INCLUDE_SPARK_PROTOBUF_JAR=false
-        export PROTOBUF_JARS_AVAILABLE=false
     fi
 
     # ALL_JARS includes dist.jar integration-test.jar avro.jar parquet.jar protobuf.jar if they exist
     # Remove non-existing paths and canonicalize the paths including get rid of links and `..`
-    ALL_JARS=$(readlink -e $PLUGIN_JAR $TEST_JARS $AVRO_JARS $PARQUET_HADOOP_TESTS $PROTOBUF_JARS || true)
+    ALL_JARS=$(readlink -e $PLUGIN_JAR $TEST_JARS $AVRO_JARS $PARQUET_HADOOP_TESTS $PROTOBUF_EXTRA_JARS || true)
     # `:` separated jars
     ALL_JARS="${ALL_JARS//$'\n'/:}"
