+++ integration_tests/src/main/python/spark_session.py
@@ -369,3 +369,36 @@ def is_hive_available():
 
 def is_hybrid_backend_loaded():
     return _spark.conf.get("spark.rapids.sql.hybrid.loadBackend", "false") == "true"
+
+def is_protobuf_runtime_available():
+    try:
+        from pyspark.sql.protobuf.functions import from_protobuf  # noqa: F401
+    except Exception:
+        return False
+
+    jvm = _spark.sparkContext._jvm
+    loader = None
+    try:
+        loader = jvm.Thread.currentThread().getContextClassLoader()
+    except Exception:
+        pass
+
+    candidates = [
+        "org.apache.spark.sql.protobuf.functions$",
+        "org.apache.spark.sql.protobuf.functions",
+    ]
+    for cls in candidates:
+        try:
+            if loader is not None:
+                jvm.java.lang.Class.forName(cls, True, loader)
+            else:
+                jvm.java.lang.Class.forName(cls)
+            return True
+        except Exception:
+            continue
+
+    try:
+        member = jvm.org.apache.spark.sql.protobuf.functions.from_protobuf
+        return type(member).__name__ != "JavaPackage"
+    except Exception:
+        return False
